👨‍❤️‍👨🏔 Checkpoint ./src/presenters/pop-overs/collection-options-pop.jsx:867163/289 ./src/presenters/collection-editor.jsx:867163/84 ./src/presenters/pages/collection.jsx:867163/409

+++ src/presenters/pop-overs/collection-options-pop.jsx
@@ -24,7 +24,10 @@ const CollectionOptionsPop = (props) => {
   }
 
   function animateThenDeleteCollection(event) {
-    animate(event, 'slide-down', () => props.deleteCollection(props.collection.id));
+    if(!window.confirm(`Are you sure you want to delete your collection?`)){
+      return;
+    }
+    animate(event, 'slide-down', () => props.deleteCollection(props.collection.id));    
   }  
   
   return(
